# DTC.Api/Dtos/Braket/CreateBraketDto.cs
using DTC.Api.Enums;
using System.ComponentModel.DataAnnotations;

namespace DTC.Api.Dtos.Braket
{
    public class CreateBraketDto
    {
        [Range(1, int.MaxValue, ErrorMessage = "Das Turnier ist ein Pflichtfeld.")]
        public int TournamentId { get; set; }

        [Range(1, int.MaxValue, ErrorMessage = "Die Reihenfolge muss größer als 0 sein.")]
        public int Sequence { get; set; }

        [StringLength(100, ErrorMessage = "Der Name der Runde darf maximal 100 Zeichen lang sein.")]
        public string? Name { get; set; }

        [Required(ErrorMessage = "Der geplante Startzeitpunkt ist ein Pflichtfeld.")]
        public DateTimeOffset PlannedStart { get; set; }

        public DateTimeOffset? PlannedEnd { get; set; }

        public BraketStatus Status { get; set; } = BraketStatus.Scheduled;

        public BraketPhase Phase { get; set; } = BraketPhase.GroupStage;
    }
}